requirements.txt
pandas==2.1.4
tabulate==0.9.0
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.2
